CR | Make Constants and MyApi files
